feat(crud-endereco-cliente)

+++ src/main/java/br/com/hellobank/api/dao/SaqueDAO.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-
-import br.com.hellobank.api.model.entidades.Saque;
+import br.com.hellobank.api.models.entidades.Saque;
 
 @Repository
-public interface SaqueDAO extends JpaRepository<Saque, Integer> {
+public interface SaqueDAO extends JpaRepository<Saque, Long> {
 
-    @Query(value = "SELECT * FROM saque WHERE conta_id = ?1", name ="findSaque", nativeQuery = true)
-    List<Saque> getSaqueById_conta(Integer contaId);
+    @Query(value = "SELECT * FROM saque WHERE conta_id = ?1", name =
+            "findSaque", nativeQuery = true)
+    List<Saque> getSaqueById_conta(Long contaId);
 }
